# finlabs_intelligence/api/openapi.json
{
  "components": {
    "schemas": {
      "AnnotationRequest": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "final_value": {
            "title": "Final Value",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "sample_id": {
            "title": "Sample Id",
            "type": "string"
          }
        },
        "required": [
          "sample_id",
          "final_value",
          "actor"
        ],
        "title": "AnnotationRequest",
        "type": "object"
      },
...
      "ChartResponse": {
        "properties": {
          "candles": {
            "items": {
              "$ref": "#/components/schemas/CandleResponse"
            },
            "title": "Candles",
            "type": "array"
          },
          "identity_checksum": {
            "title": "Identity Checksum",
            "type": "string"
          },
          "interval": {
            "default": "1min",
            "title": "Interval",
            "type": "string"
          },
          "market": {
            "title": "Market",
            "type": "string"
          },
          "replayable": {
            "default": false,
            "title": "Replayable",
            "type": "boolean"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "market",
          "symbol",
          "identity_checksum",
          "candles"
        ],
        "title": "ChartResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local-first collection and labeling API.",
    "title": "FinLabs News Intelligence API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/annotations": {
      "post": {
        "operationId": "createAnnotation",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnnotationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnotationRevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Annotation"
      }
    },
    "/api/annotations/{sample_id}": {
      "get": {
        "operationId": "getAnnotationHistory",
        "parameters": [
          {
            "in": "path",
            "name": "sample_id",
            "required": true,
            "schema": {
              "title": "Sample Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnnotationRevisionResponse"
                  },
                  "title": "Response Getannotationhistory",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Annotation History"
      }
    },
    "/api/catalog": {
      "get": {
        "operationId": "searchCatalog",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogItemResponse"
                  },
                  "title": "Response Searchcatalog",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Catalog Search"
      }
    },
    "/api/charts/{symbol}": {
      "get": {
        "operationId": "loadChart",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "market",
            "required": true,
            "schema": {
              "pattern": "^(KOSPI|KOSDAQ)$",
              "title": "Market",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "start_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Start At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "end_at",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "End At",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "chart_type",
            "required": false,
            "schema": {
              "default": "minute",
              "pattern": "^(minute|daily)$",
              "title": "Chart Type",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval_minutes",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Interval Minutes",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Chart Load"
      }
    },
    "/api/datasets": {
      "post": {
        "operationId": "freezeDataset",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreezeDatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Dataset Route"
      }
    },
    "/api/datasets/preview": {
      "post": {
        "operationId": "previewReaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReactionPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReactionPreviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Reaction Route"
      }
    },
    "/api/datasets/{dataset_id}/exports": {
      "get": {
        "operationId": "getDatasetExports",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Status Route"
      },
      "post": {
        "operationId": "exportDataset",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Dataset Route"
      }
    },
    "/api/datasets/{dataset_id}/exports/{sink}/retry": {
      "post": {
        "operationId": "retryDatasetExport",
        "parameters": [
          {
            "in": "path",
            "name": "dataset_id",
            "required": true,
            "schema": {
              "title": "Dataset Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sink",
            "required": true,
            "schema": {
              "title": "Sink",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDatasetResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Export Route"
      }
    },
...
    "/api/news/discover": {
      "post": {
        "operationId": "discoverNews",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiscoverNewsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscoverNewsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover News"
      }
    },
...
  }
}